cleanup: use resolveSubpath helper to resolve all subdirectory paths
with an updated comment

SOME ADDITIONAL, GENERAL NOTES:

MOTIVATION: The primary motivation of using `path.resolve()` rather
than `path.join()` is to enable the tests to check the results in a
less hacky manner.

HOW: This was accomplished by using const helper definitions to
facilitate the replacement of `process.cwd()` & `path.join`
in straightforward manner, with a relatively low risk.